+++ src/commands/hardis/org/monitor/backup.ts
@@ -9,7 +9,7 @@ import { buildOrgManifest } from "../../../../common/utils/deployUtils";
 import { execCommand, filterPackageXml, uxLog } from "../../../../common/utils";
 import { MetadataUtils } from "../../../../common/metadata-utils";
 import { CONSTANTS } from "../../../../config";
-import { NotifProvider } from "../../../../common/notifProvider";
+import { NotifProvider, NotifSeverity } from "../../../../common/notifProvider";
 import { MessageAttachment } from "@slack/web-api";
 import { getNotificationButtons, getOrgMarkdown } from "../../../../common/utils/notifUtils";
 import { generateCsvFile, generateReportPath } from "../../../../common/utils/filesUtils";
@@ -183,11 +183,16 @@ You can remove more metadata types from backup, especially in case you have too 
       this.outputFilesRes = await generateCsvFile(this.diffFilesSimplified, this.outputFile);
     }
 
-    // Send notifications
+    // Build notifications
+    const orgMarkdown = await getOrgMarkdown(this.org?.getConnection()?.instanceUrl);
+    const notifButtons = await getNotificationButtons();
+    let notifSeverity: NotifSeverity = "log";
+    let notifText = `No updates detected in ${orgMarkdown}`
+    let notifAttachments: MessageAttachment[] = [];
     if (this.diffFiles.length > 0) {
-      const orgMarkdown = await getOrgMarkdown(this.org?.getConnection()?.instanceUrl);
-      const notifButtons = await getNotificationButtons();
-      const attachments: MessageAttachment[] = [
+      notifSeverity = "info";
+      notifText = `Updates detected in ${orgMarkdown}`;
+      notifAttachments = [
         {
           text: this.diffFiles
             .map((diffFile) => {
@@ -201,21 +206,23 @@ You can remove more metadata types from backup, especially in case you have too 
             .join("\n"),
         },
       ];
-      globalThis.jsForceConn = this?.org?.getConnection(); // Required for some notifications providers like Email
-      NotifProvider.postNotifications({
-        type: "BACKUP",
-        text: `Updates detected in ${orgMarkdown}`,
-        buttons: notifButtons,
-        attachments: attachments,
-        severity: "info",
-        sideImage: "backup",
-        attachedFiles: this.outputFilesRes.xlsxFile ? [this.outputFilesRes.xlsxFile] : [],
-        logElements: this.diffFilesSimplified,
-        metric: this.diffFilesSimplified.length
-      });
     } else {
       uxLog(this, c.grey("No updated metadata for today's backup :)"));
     }
+
+    // Post notifications
+    globalThis.jsForceConn = this?.org?.getConnection(); // Required for some notifications providers like Email
+    NotifProvider.postNotifications({
+      type: "BACKUP",
+      text: notifText,
+      buttons: notifButtons,
+      attachments: notifAttachments,
+      severity: notifSeverity,
+      sideImage: "backup",
+      attachedFiles: this.outputFilesRes.xlsxFile ? [this.outputFilesRes.xlsxFile] : [],
+      logElements: this.diffFilesSimplified,
+      metric: this.diffFilesSimplified.length
+    });
 
     return { outputString: "BackUp processed on org " + this.org.getConnection().instanceUrl };
   }